Remove group_id and use the real ID instead
Also add helper set_admin_email

+++ models/migrate.go
@@ -33,7 +33,7 @@ func Migrate() {
 	for rows.Next() {
 		aNote := Note{}
 		aNote.AuthorID = int64(1)
-		aNote.GroupID = int8(1)
+		aNote.GroupID = int64(1)
 		var nContent, nURL, nFlags sql.NullString
 		var dateLog, timeStamp sql.NullInt64
 		if err := rows.Scan(
